Split up the parsing compiler into separate type structs.

+++ pkg/protobuf/compiler/src/spec.rs
@@ -9,7 +9,7 @@ use protobuf_descriptor as pb;
 #[derive(Clone, Debug)]
 pub enum Constant {
     Identifier(String),
-    Integer(isize),
+    Integer(i64),
     Float(f64),
     String(Vec<u8>),
     Bool(bool),
